+++ packages/components-extra/src/components/Navbar/Navbar.tsx
@@ -1,5 +1,4 @@
 import React, { forwardRef } from 'react'
-import PropTypes from 'prop-types'
 import styled from 'styled-components'
 import { AppBar, Toolbar } from '@material-ui/core'
 
@@ -11,23 +10,13 @@ import MenuItem from './components/MenuItem'
 
 import { NavbarProps, NavbarType } from './types'
 
-const Navbar = forwardRef(
-  ({ children, isSticky = false, ...rest }: NavbarProps, ref: React.Ref<HTMLElement>) => {
-    return (
-      <AppBar position={isSticky ? 'sticky' : 'static'} ref={ref} {...rest}>
-        <Toolbar>{children}</Toolbar>
-      </AppBar>
-    )
-  },
-) as NavbarType
-
-Navbar.propTypes = {
-  /**
-   * Set to true if you want the Navbar to stick to its initial position. It will give the
-   * 'position: sticky' css rule to the main container.
-   */
-  isSticky: PropTypes.bool,
-}
+const Navbar = forwardRef(({ children, ...rest }: NavbarProps, ref: React.Ref<HTMLElement>) => {
+  return (
+    <AppBar ref={ref} {...rest}>
+      <Toolbar>{children}</Toolbar>
+    </AppBar>
+  )
+}) as NavbarType
 
 /**
  * Exposed components.