+++ packages/components-extra/src/index.ts
@@ -1,4 +1,5 @@
 export * from './components'
 export * from './theme'
+export * from './hooks'
 
 export { default as resetCSS } from './utils/resetCSS'