+++ packages/docs/src/components/Layout.tsx
@@ -1,20 +1,24 @@
 import React from 'react'
-import { createGlobalStyle } from 'styled-components'
+import styled, { createGlobalStyle } from 'styled-components'
 import { StyledProvider, BackToTop } from 'components-extra'
 
-import { ModeProvider } from '@theme'
+import { ModeProvider, theme } from '@theme'
 
 import Header from './Header'
 import Navigation from './Navigation'
 
 const GlobalStyle = createGlobalStyle`
-  ${({ theme }): string => `
+  ${({ theme: { palette, transitions } }): string => `
     html {
-      background: ${theme.palette.background.default};
-      scrollbar-color: #bfbfbf ${theme.palette.background.default};
-      transition: color 0.2s ease-out, background 0.2s ease-out;
+      background: ${palette.background.paper};
+      scrollbar-color: #bfbfbf ${palette.background.default};
+      transition: ${transitions.backgroundColor};
     }
   `};
+`
+
+const MainContainer = styled.main`
+  padding-top: 84px;
 `
 
 const Layout: React.FC<LayoutProps> = ({ children }: LayoutProps) => {
@@ -22,20 +26,22 @@ const Layout: React.FC<LayoutProps> = ({ children }: LayoutProps) => {
 
   return (
     <ModeProvider dark={dark} setDark={setDark}>
-      <StyledProvider dark={dark}>
+      <StyledProvider dark={dark} theme={theme}>
         <GlobalStyle />
         <Header />
-        <Navigation />
-        <div
-          style={{
-            margin: `0 auto`,
-            maxWidth: 960,
-            height: '1280px',
-            padding: `0 1.0875rem 1.45rem`,
-          }}
-        >
-          <main>{children}</main>
-        </div>
+        <MainContainer>
+          <Navigation />
+          <div
+            style={{
+              margin: `0 auto`,
+              maxWidth: 960,
+              height: '1280px',
+              padding: `0 1.0875rem 1.45rem`,
+            }}
+          >
+            {children}
+          </div>
+        </MainContainer>
         <BackToTop />
       </StyledProvider>
     </ModeProvider>