+++ packages/docs/src/icons/index.ts
@@ -1,3 +1,4 @@
 export { default as Dark } from './Dark'
 export { default as Github } from './Github'
 export { default as Light } from './Light'
+export { default as Menu } from './Menu'