+++ packages/docs/src/theme/index.ts
@@ -1,1 +1,2 @@
+export { default as theme } from './theme'
 export * from './ModeContext'