+++ packages/docs/src/theme/theme.ts
@@ -1,0 +1,5 @@
+export default {
+  transitions: {
+    backgroundColor: 'color 0.2s ease-out, background 0.2s ease-out',
+  },
+}